Cleaned API Structure Added Redis Logic (Cache)
