+++ app/api/bus-operations/[BusAssignmentID]/route.ts
@@ -7,49 +7,6 @@ enum BusOperationStatus {
   Completed = 'Completed',
 }
 
-type BusAssignmentWhereInput = {
-  IsDeleted?: boolean;
-  Status?: BusOperationStatus;
-};
-
-export async function GET(request: Request) {
-  try {
-    const url = new URL(request.url);
-    const status = url.searchParams.get('status');
-
-    const whereClause: BusAssignmentWhereInput = { IsDeleted: false };
-
-    if (status !== null) {
-      if (
-        status === BusOperationStatus.NotStarted ||
-        status === BusOperationStatus.InOperation ||
-        status === BusOperationStatus.Completed
-      ) {
-        whereClause.Status = status;
-      } else {
-        return NextResponse.json({ error: 'Invalid status value' }, { status: 400 });
-      }
-    }
-
-    const busAssignments = await prisma.busAssignment.findMany({
-      where: whereClause,
-      include: {
-        RegularBusAssignment: true,
-        TicketBusAssignments: {
-          include: {
-            TicketType: true,
-          },
-        },
-      },
-    });
-
-    return NextResponse.json(busAssignments);
-  } catch (error) {
-    console.error('Failed to fetch bus assignments:', error);
-    return NextResponse.json({ error: 'Failed to fetch bus assignments' }, { status: 500 });
-  }
-}
-
 type BusAssignmentUpdateData = Partial<{
   Status: BusOperationStatus;
   Battery: boolean;
