+++ package.json
@@ -14,6 +14,7 @@
     "@hello-pangea/dnd": "^18.0.1",
     "@prisma/client": "^6.6.0",
     "@supabase/supabase-js": "^2.49.4",
+    "@upstash/redis": "^1.34.9",
     "bootstrap": "^5.3.6",
     "cuid": "^3.0.0",
     "next": "15.3.0",
